Use postgresql_query for primary_check
We lazily relied on «repmgr cluster show» to exit without error if the
primary was registered, but were duly punished for our sloppiness. If
the primary was already registered, but «repmgr cluster show» exited
with an error because it could not, say, connect to the witness, we
would try to re-register the primary and fail.