Separate bdr_group_create/join from _wait_for_ready
We were calling both in the same transaction, and the wait hung forever
for reasons that are not very clear. But it made debugging harder, since
bdr_nodes and similar metadata tables were empty, since the originating
transaction had not committed yet. Separating the two will let us see
what's going on in a bit more detail.

Also, it's not necessarily safe to call bdr_node_join_wait_for_ready()
concurrently on multiple nodes, though it may work on BDRv2. So we run
it one by one, the same way we sequence the group joins.