Revise local known_hosts generation
1. We no longer generate entries for instances with platform=='bare'
2. We generate entries into $cluster/tpa_known_hosts (instead of
   known_hosts). We will not touch known_hosts if it exists.
3. We set UserKnownHostsFile to "known_hosts tpa_known_hosts"

Now if you use platform=bare and use ssh-keyscan to populate known_hosts
with entries for the bare instances, it will work, and running provision
again will not overwrite your custom entries.

This still leaves the problem of generating /etc/ssh/ssh_known_hosts on
the instances, but it makes the situation no worse than it already is,
and we can deal with it separately.