Set homehost to 'any' when creating or assembling RAID arrays
The mdadm manpage says (about the ``--homehost=`` option):

    «The special name "any" can be used as a wild card.  If an array is
    created with --homehost=any then the name "any" will be stored in
    the array and it can  be  assembled  in the same way on any host.
    If an array is assembled with this option, then the homehost
    recorded on the array will be ignored.»

We specify ``--homehost=any`` when creating or assembling RAID arrays,
so that the array will be automatically assembled on boot as /dev/md0
(or whatever the expected name is) and not as /dev/md/othername:0 to
avoid conflicts.

This means that an array that was created on instance x can be assembled
the same way on a new instance y (== rehydration), or on instance x with
a different hostname (== reboot, see below).

Volumes created by clusters prior to this change will still work fine on
rehydration because "any" will cause reassembly to ignore whatever name
is already set for the volumes (i.e., the hostname of the instance that
originally created them). New volumes will be created with "any".

We also revert 938453c, which took a different approach to the problem
by trying to ensure that the hostname was set to what we expect before
the array was assembled. This appeared to work, but had some problems:

1. If you provisioned an instance with a RAID volume and rebooted it
   (i.e., without deploying), the hostname would not be set the second
   time (echo foo > /proc/sys/kernel/hostname isn't persistent, and the
   user-data isn't run again) and the RAID array would not be assembled
   as /dev/md0

2. Even if you *did* deploy before rebooting, setting the hostname via
   /etc/hostname seems to happen after the RAID array is auto-assembled
   under a different name (because of the HOMEHOST mismatch), and things
   would be even worse—sys/volumes would have added an fstab entry for a
   filesystem built on the RAID array, and the system would never finish
   booting because it couldn't mount that filesystem.

3. Setting the hostname early broke "hostname -f", which is used by the
   2ndQ repository registration script, so we had to add a temporary
   /etc/hosts entry to enable the repositories (so now we can revert
   41464a6 too).