Create links from /dev/{sd,xvd}* to /dev/nvme*n1
https://docs.aws.amazon.com/AWSEC2/latest/UserGuide/nvme-ebs-volumes.html
says:

    «With the following instances, EBS volumes are exposed as NVMe block
    devices: C5, C5d, i3.metal, M5, M5d, R5, R5d, T3, and z1d. The
    device names are /dev/nvme0n1, /dev/nvme1n1, and so on. The device
    names that you specify in a block device mapping are renamed using
    NVMe device names (/dev/nvme[0-26]n1).»

I had hoped that one could specify /dev/nvme1n1 and similar in the block
device mapping when creating the request, and just translate xvd* to the
right nvmeXn1 name based on the instance type. But in fact you can't ask
for an nvme device. You have to use sdX/xvdX names as before, and it's
up to AWS to decide whether you get nvme volumes or something else (the
number of nvme interfaces available depends on the instance type too).
The name is also not stable (i.e., vol-xxx may be attached as nvme1n1
when the instance is created, and nvme2n1 after it's rebooted).

The only way to figure out which volume in the block device mapping a
particular nvme device corresponds to is to install nvme-cli and read
some memory mapped from the nvme controller:

    [ec2-user ~]$ sudo nvme id-ctrl -v /dev/nvme1n1
    NVME Identify Controller:
    vid     : 0x1d0f
    ssvid   : 0x1d0f
    sn      : vol01234567890abcdef
    mn      : Amazon Elastic Block Store
    ...
    0000: 2f 64 65 76 2f 73 64 6a 20 20 20 20 20 20 20 20 "/dev/sdf..."

https://github.com/oogali/ebs-automatic-nvme-mapping has a script to
discover the requested volume name and create suitable symlinks, plus
udev rules to trigger it on nvme* devices.

(Amazon Linux creates these links automatically and provides an
ebsnvme-id command that prints the right volume name too.)

Unfortunately, on the t3.micro instance I tested with, (a) what showed
up in the nvme controller memory was "sdf" or "xvdb", not "/dev/xxx", so
I had to modify the script a bit, and (b) udev didn't trigger the rule
after it was added from user-data, so I had to execute the script from
user-data as well. With those changes, we get /dev/sdf → /dev/nvme1n1
symlinks, and we don't have to worry about translation.

Many thanks to Craig A. for doing all the detective work to arrive at
this solution.

Fixes #259