Separate group-join tasks into separate include file
With BDR3, the bdr.join_node_group() functions waits until the join has
completed and the local node is ready. With BDR2, however, we need to
call bdr.bdr_node_join_wait_for_ready() before moving on to the next
node, but we need to do so in a separate transaction from where we
invoke bdr.bdr_group_join().