Remove pg_visibility for 9.5 too
This change is a hack—we should annotate extensions with conditions and
have a task that removes any extensions that don't meet their condition
based on postgres_version. But right now the problem affects only this
extension, so we'll defer the "proper fix" until we have another case.

Thanks to Boriss for reporting the problem.