Permit - in cluster_name
We disallowed this at one time because we used the name in repmgr.conf,
but we don't do that any more, and enough people have tried to create
clusters named "foo-bar" that it seems worth changing.

If we use the value in a context where - is inconvenient, e.g., as a
database name (cf. default_bdr_database), we can convert - to _ anyway;
this is how we should handle any unforeseen problems that may arise
after this commit.