Set a default archive_timeout of 6h
Not setting archive_timeout at all means that a server with zero traffic
will never switch WAL segments. That has unpleasant consequences for any
Barman backups we take from a replica, because the WAL segment required
to complete the backup will never make it from xxx.partial to wals/

(If there are other valid backups already, you can recover from this at
the expense of having to use an older base backup and restore more WAL.)

We don't really want to wait for 6h for a backup to become valid, but
this establishes a definite backstop for the extreme corner case. Note
that even if we are not doing anything, checkpoint_timeout will ensure
that *some* WAL gets written at least once a day.