Support volatile subscriptions via --use-volatile-subscriptions
We normally configure tpa_2q_repositories with persistent subscriptions,
but for test clusters that we know will be short-lived, we can now run
tpaexec configure with --use-volatile-subscriptions to enable volatile
subscriptions instead. These will expire after 24h.

Since we don't know whether the subscription has expired or not, we just
always re-subscribe to the repositories on every deploy. This means that
we'll create a lot of volatile subscriptions for repeated deploys, but
they will all eventually expire.

For a cluster where we somehow know in advance that we will only ever
run deployment once, it's just what the doctor ordered. It's up to you
to decide whether you would be better off with a persistent subscription
or a volatile subscription (or subscriptions).