Add "tags: always" to every include_{tasks,role}
To make --tags x work, we need to be sure that any task tagged x is
visible during execution, which means that dynamic includes must always
be executed, so that the included tasks can be evaluated against the
tags specified. This is also important for --skip-tags x (which would
not work even with this code, but for the change in 2ndQuadrant/ansible
to not skip always-tagged tasks).

In future, some roles may choose to exempt certain "internal" includes
from the always-execute rule, but the way it is now is the safe way to
do it, where specifically-tagged tasks can be included or excluded
without regard to the structure of inclusions.