Don't enforce cluster_tags[Owner] for non-aws clusters
Without this change, deprovision would fail if you provision a cluster
as admin (or one of the other proscribed usernames) with «--owner xxx»
to satisfy the check.

We really care about the owner being set sensibly for EC2 instances (or
any other cloud instances that we have to pay for). It doesn't matter
for local Docker and Vagrant instances at all.

Even with this change, deprovision will fail for EC2 clusters where the
owner was set at provision time (but setting the name at configure time
would solve the problem). We can worry about that problem later.