Use new «meta: reload_vars» to pick up xxx_password.yml definitions
Depends on the feature/reload_vars branch in 2ndQuadrant/ansible.

We want to load a new group_vars entry created dynamically by the
playbook, and there's (surprisingly) no way to do that.

We used to include_vars the newly-generated file (aside: xxx_password
would be therefore be defined as a host var on the first run, and a
group var on subsequent runs when loaded from the inventory), but we
switched from meta dependencies to include_role, which meant that the
include_vars was no longer run on every host (not that it was exactly
reliable before), so it broke.

Now we depend on reload_vars to pick up the newly-created file.