Switch to taking backups from the primary
[After much research…] There are only two safe ways to configure backups
with Barman at the moment:

1. Use archive_command on the primary and take backups from either the
   primary or a replica

2. Stream from the primary and take backups from the primary

Unfortunately, streaming from a replica and taking backups from the same
replica (which is what we did earlier) does not work correctly, even if
we forced a WAL switch before and after the initial backups.

Also, streaming from the primary and taking backups from a replica ought
to work, but is not currently supported by Barman (barman check uses the
conninfo to check for the existence of a replication slot, so it breaks
if you point streaming_conninfo to another server).

Until Barman can be modified to work better in the (desirable) case of
streaming+backups from a replica, we should default to the safe setup.

(This change only affects new clusters. One way to fix existing clusters
would be to set archive_command on the primary and leave everything else
alone (at the expense of 2x WAL transfer, which may or may not matter in
a given setup). The more complete fix would be to do that, but also stop
streaming from the replica. That can be implemented in a later commit.)