Force a WAL switch on the primary after first-backups from a replica
When taking backups from a primary, Barman will pg_switch_wal() after
the backup completes. It doesn't know how to do that when taking a
backup from a replica, so we help it along.

Note that this doesn't make it safe to stream from a replica and also
take backups from it—there's still the problem with Barman seemingly
misdetecting the start LSN of the Backup, as described in RM#6429.

On the other hand, if you're using archive_command on the primary and
taking backups from a replica, this will correctly finalise the backup
when barman cron archives the new segment (WAITING_FOR_WALS→DONE).