Add "sys" tests (only a package list for now)
We introduce test/tasks/sys.yml that can perform system-level tasks at
the beginning of the test run, and os/*/* include files that handle any
distribution-specific tasks.

This is the right place to collect other system-level information about
the cluster (lshw? lsblk?), but at the moment we only collect a list of
all installed packages.

Delivered: #6418