Add --replicate-deletes-locally option to CAMO2x2
This option causes three additional replication sets (all with
autoadd_tables enabled) to be created:

    1. cluster_global, which replicates insert/update but not
       delete/truncate
    2. local_{a,b}_deletes, which replicate only delete/truncate but not
       insert/update.

The nodes in location a subscribe to {cluster_global,local_a_deletes}
and likewise for those in b. The result is that insert/update are
replicated between locations, but deletes remain confined to the
location they were issued in.