Eliminate duplication between READMEs and docs/
It was becoming difficult to keep {platforms,architectures}/*/README.md
and docs/{platform,architecture}-*.md in sync, so we just symlink the
former to the latter to make maintenance easier.

We should keep in mind that the .md will be read directly on-screen
(e.g., via ``tpaexec info architectures/x``) and write accordingly.