Support building some extensions from source
Now you can do this:

    instances:
    - …
      vars:
        install_from_source:
        - name: pglogical
          git_repository_url: https://whatever
          git_repository_ref: master
        - name: bdr
          git_repository_url: https://whatever
          git_repository_ref: REL_SOMETHING
          build_commands:
          - git clean -dxf
          - make
          - make install
          build_environment:
            FOO: 42

The name, git_repository_url, and git_repository_ref are required. The
build_commands default to make and make install; build_environment is
optional. The mechanism is general enough to build anything—the only
thing we do to help Postgres extensions is to make sure the right
pg_config is in the PATH when running (all of) the build_commands.
Build output is appended to /tmp/build-xxx.log for debugging.

(Of course, all of this could go into instance_defaults as well.)

A few considerations:

1. You have to make sure the target instances can access the repository,
   e.g., by (separately) configuring deploy keys, or using ssh agent
   forwarding (for which we have special support), or something else.

2. If 'name: pglogical' is in the list, then postgres/pkg won't try to
   install pglogical packages. Likewise, if 'name: bdr' is specified, it
   won't install BDR packages. This mechanism is not version-specific,
   and does not know about the dependency of bdr3 on pglogical3. So if
   you put pglogical into install_from_source but leave bdr3 out, bad
   things will happen.

3. pglogical and bdr are the only extensions that we currently install
   packages for by default, so they're the only ones we know how to
   treat specially (in the sense of not installing the packages if they
   are mentioned in install_from_source).

4. Any packages that are required to build the sources will have to be
   installed via --extra-postgres-packages or by some other means.

I couldn't think of a much better name than "install_from_source", other
than monstrosities like "extensions_to_be_installed_from_source", which
is too long and perhaps a little too specific.

In terms of implementation, we introduce a new postgres/install role
that invokes postgres/<src|pkg> and then invokes a new src/install role
in a loop over install_from_source[] list. src/install knows only how to
clone a repository and run a bunch of commands; we could add more smarts
to it later (e.g., building from source tarballs, separating the source
and build directories, and so on).