Set changed_when for restorecon
We still always run «restorecon -R -v $PGDATA» when selinux is enabled,
because there's no really good way to figure out whether it would change
anything other than running with -n and repeating the task without -n if
there's any output. But if there's no output from «-R -v», we know that
nothing was changed, so we can make the output green.

(A better way to do this might be to use -D along with -R, which should
make the command run faster. But that's a larger change the requires a
fair bit of testing. This change is purely cosmetic, to avoid yellow
output when nothing's actually changed.)

Delivered: #6054