Add proper wildcard support to s3_uploads
When the s3_uploads setting was originally implemented, the suggested
syntax for wildcards was as follows

    s3_uploads:
    - foo
    - "{{ query('fileglob', 'id_*') }}"

Unfortunately, this didn't actually work because cwd wasn't cluster_dir,
and fileglob's "dwimmed_path" search didn't do anything helpful. Instead
of making a custom filter to address this, we now accept just "id_*" as
an entry and call fileglob with the right base directory to expand the
list of files to upload. So the recommended syntax is now:

    s3_uploads:
    - foo
    - id_*
    - /etc/hosts

It is not currently an error for an entry to not exist.

Based on a proposal and patch from Craig A.