Update ephemeral_storage table
We update the table of instance types and the corresponding number of
instance store volumes based on the information at
https://docs.aws.amazon.com/AWSEC2/latest/UserGuide/InstanceStorage.html#instance-store-volumes

We exclude NVMe SSD instance store volumes, because we can't reliably
map them to the specified device name—they will show up as /dev/nvmeX
and ``nvme id-ctrl`` doesn't return a device name like 'xvdc' that we
can use to create a symlink to the expected name (as we normally do in
user-data for other volumes attached via NVMe). Since the nvmeX names
aren't stable (and cannot be specified in block device mappings anyway),
we can't support ephemeralX annotations for NVMe volumes.

Delivered: #6053