Identify BDR3 logical standby instances as [bdr,primary,readonly]
Earlier, the only difference between a [bdr,primary] and a logical
standby was the value of "bdr_node_pause_in_standby".

Since then, we've felt the need to identify/exclude standby instances
from various r/w operations (e.g., group creation, pgbench -i), so it is
now necessary to assign them a distinct role for easier identification.

We choose [bdr,primary,readonly] because a logical standby is like a BDR
primary in how it's initialised (through the BDR group join mechanism),
not like a [bdr,replica] that's initialised from a pg_basebackup. The
'readonly' designation is more of a description of the intended function
of the logical standby rather than a literal one, since the instance can
have unreplicated writes.

To maintain backwards compatibility with existing clusters that have
instances with bdr_node_pause_in_standby set, we add 'readonly' to the
role for such instances, and later set bdr_node_pause_in_standby to true
if the 'readonly' role is set.

Finally, we exclude readonly instances from the determination of which
instance to create the bdr_node_group on.

[Aside: the assignment of a first_primary could be hoisted from bdr3.yml
to main.yml and the designation used in bdr2.yml as well, but that would
be a cosmetic change, so I avoided it in the interest of minimalism.]