Add an --os-image configure option to override default image selection
This is required to override the forced selection of centos/systemd for
Docker when alternative images are available. It's implemented as an
architecture option rather than being specific to the Docker platform,
but left undocumented to avoid confusion.

If you use this option, you are responsible for providing suitable
images that meet deployment needs. Things will break otherwise.