Avoid conditional inclusion failure
If you run deploy with ``--tags bdr``, this test would fail because the
tasks that compute and set the value aren't run; that's fine, we can
skip the include in that case.

In general, any conditional include_{role,tasks} (tagged with always)
should be written to avoid failure if the condition depends on some
variable that isn't set because of the selected combination of tags.
(This commit doesn't look for other such cases, though.)