Disable repmgr tests for bdr instances
Repmgr considers the (mostly inactive) BDR group slot to be a CRITICAL
error, on the basis that any inactive slot must be bad and wrong (not so
for BDR3 group slot, which is advanced intermittently by the code). This
leads to the repmgr node check always failing in a BDR cluster.

It's inconvenient (and overbroad) to «--skip-tags fail» every time, so
this change allows the checks to continue on non-BDR instances while
temporarily excluding BDR ones. Once repmgr learns to ignore the group
slot, this change can be reversed.