Exclude both readonly and replica BDR instances from running pgbench -i
Unlike postgres/bdr, which is applied only to [bdr,primary] instances,
the code in pgbench is also applied to replicas. If the code happened to
select a replica as "first_primary", then it would end up not executing
pgbench -i anywhere because the condition excluded replicas separately.

Instead of further complicating the host selection expression, we add a
new filter to return the members of a group that are not in a list of
other groups. For example,

    groups|members_of(bdr_node_group, not_in=['role_readonly','role_replica'])

now returns the list of r/w [bdr,primary] instances.