Enable TLS for pgbouncer and haproxy (in all directions)
This is the default TLS-everywhere version. In future, we can consider
extending this to allow TLS to be turned off in specific cases (e.g.,
when pgbouncer and haproxy are on the same instance, or pgbouncer and
postgres are on the same instance).

Development and testing by Niccolò and Anna; thanks also to Jonathan for
analysis. Committed with a couple of minor changes omitted (the *_user
change was applied separately, the pgsql check is required, and the
pgbouncer path change should be considered separately (the path is
correct on RH)).

Delivered: #6055

Signed-off-by: Anna Bellandi <anna.bellandi@2ndquadrant.com>
Signed-off-by: Niccolò Fei <niccolo.fei@2ndquadrant.com>
Signed-off-by: Abhijit Menon-Sen <ams@2ndQuadrant.com>