Raise bdr.global_lock_timeout to 600s for pgbench -i
We've experienced intermittent DDL lock acquisition timeouts during the
last stage of the pgbench -i run (i.e., creating primary keys). It's not
deterministic, but we hope it's just slow rather than a bug, and that
temporarily raising the timeout will make it succeed.

This is just a quick hack: if we need anything more of this nature, we
should probably create a dedicated pgbench user and set these using
«ALTER USER … SET …» instead of ALTER SYSTEM.

References RM#7146