Always run pgbench -i
Since pgbench behaves sensibly if you rerun pgbench -i, checking for
pgbench_history and avoiding running pgbench -i buys us nothing; that
test is so simplistic that we would avoid running pgbench -i even if it
did not complete successfully the first time, so long as it managed to
create the table (_one_ table).

Besides, running pgbench -i has revealed enough problems to be worth the
expense, and the only available workaround to make it run every time
(dropping the tables by hand in between runs) is tedious.