Implement «--enable-camo» as an option for all BDR architectures
This is quite a blunt hammer: it just takes [bdr,primary] instances that
don't have bdr_node_camo_partner set, pairs them up by index, and sets
them as each other's CAMO partner. If there are any instances leftover,
so be it. It's not very sophisticated, but it is enough to experiment
with BDR-Simple and CAMO.

CAMO2x2 does not expose --enable-camo (for obvious reasons).

Fixes RM#7053