Don't install dl/default/release if tpa_2q_repositories is explicitly []
We could set yum_repository_list to [] to suppress the installation of
all external repositories, but we would separately still try to install
dl/default/release; in cases where we don't want one, we most likely do
not want the other as well.

Now we can set tpa_2q_repositories to [] to avoid that (in addition to
using «--skip-tags public_repos» as a less convenient workaround).

Fixes #7173