Add an 'empty' test and simplify some expressions
Cosmetic.

We can write 'x is empty' instead of 'x == []' or 'x|length == 0', and
'x is not empty' instead of 'x != []' or 'x|length > 0' wherever it's
clearest to do so.

This is especially useful when writing role expressions like

    role|intersect(['primary','replica', …]) is not empty

instead of a long "'primary' in role or 'replica' in role or …"
expression.