Correctly set changed_when for ``find … -exec rm``
We need to remove host_vars files in two places (to get rid of the old
0[23]*.yml files, and during deprovisioning), and there could be a lot
of such files, so we want to do it quickly. Now we use a similar block
of code in both places, and make sure the exit status reflects whether
the block did something or not.

Also make sure args→executable is set to /bin/bash for all shell module
invocations within platforms/.