Add 'region' to export_if_set in export_vars
We used to always_export region (for some reason that was not explained
properly in the original commit message), but nothing under roles/ used
it at all. On the off-chance that someone's hook scripts depended on it,
we add it back to export_if_set here to avoid compatibility problems.

Still, new code should not depend on it, and should almost certainly use
either location (platform-independent) or ec2_region (AWS-specific)
instead.