Revise ``tpaexec update-postgres`` to update all Postgres-related packages
The existing M1 ``update-postgres`` command worked fine, but it poorly
duplicated some of the logic in postgres/pkg, and would have had to do a
lot more of the same to update additional packages or work for other
architectures. Instead, we

1. Introduce a postgres/update role, analogous to postgres/install,
   which is responsible for doing whatever is necessary to update all
   Postgres-related packages based on postgres_installation_method.

2. Introduce postgres/pkg/tasks/update.yml, and have both main.yml and
   update.yml include list-packages.yml, which enumerates all packages
   that need to be installed (or updated), so that all of the logic is
   in one place.

3. Introduce commands/update-postgres.yml for the BDR-Simple,
   BDR-Always-ON, and CAMO2x2 architectures. These are similar to the M1
   version, but don't need to do any switchovers, just apply the updates
   instance by instance (with ``serial: 1``).

Note that it's currently possible only to update to the latest available
version of all relevant packages, not to specify arbitrary versions.

If you run update-postgres on a cluster that has Postgres installed from
source (i.e., postgres_installation_method == 'src') or has extensions
listed in install_from_source, it will fail with an error. When we have
time to test the process, it should be quite simple to make it work for
source-installed clusters too (since all it needs to do is repeat the
installation process). Meanwhile, redeploying should work on such
clusters.

Signed-off-by: Abhijit Menon-Sen <ams@2ndQuadrant.com>