Restart postgres if the package list was changed
Now that we're updating packages other than Postgres itself, we can't
rely on the postgres package update causing a restart, and we can't be
sure whether a particular update requires a restart or not. On the basis
that most extensions require a restart after the update, we now restart
Postgres if any packages were updated.

Note the new symlink from postgres/pkg/handlers to postgres/handlers.
When we include_role postgres/pkg directly, it doesn't pick up the
parent role's handlers.