Omit -i/--vault-password-file in ``tpaexec playbook`` if inapplicable
We now add ``-i inventory`` to the command-line only if the inventory
exists, and ``--vault-password-file …`` only if the said file exists.

This makes it possible to use ``tpaexec playbook`` even on a cluster
that has not yet been provisioned (and so, for example, to be able to
run deprovision twice).