Add an experimental ``tpaexec eval`` command
This just does ``debug: msg="{{ expression }}"``, but it does it after
running init.yml, so that all the basic stuff we expect to be defined in
a deployment playbook is already in place, which is not the case if you
run ``tpaexec cmd -m debug …``.

You can run it like this:

    tpaexec eval . "{{ some expression }}"

    # Don't include roles/init
    tpaexec eval . "{{ some expression }}" --no-init

    # Run only on specified hosts
    tpaexec eval . "{{ some expression }}" --hosts foo,bar,baz

Expression must come first, options afterwards.