Don't require --distribution for --platform bare
The requirement doesn't seem to have been a fully thought-out decision
in the original code. It doesn't make sense to require it in this case,
since we have no control over what's installed on bare instances anyway.
Nor does it make sense to declare what to expect, since fact discovery
is expected to take care of that.