Allow platforms to set their own hostname validation patterns
We want to allow foo.example.com for bare instances, but not for AWS
instances (if you do tat, you end up with hosts named foo.example.com in
the static inventory and foo_example_com in the dynamic inventory, which
makes things confusing very fast).

The default is still a "normal" single-word hostname (^[a-z0-9-]+$), but
now bare instances may specify an FQDN instead.