Support NVME devices for both EBS and instance store volumes
We want the following volume layout on any EC2 instance:

    /dev/xvda           EBS root volume
    /dev/xvd[b-e]       instance store volumes
    /dev/xvd[f-p]       additional EBS volumes

Before NVMe devices were introduced for EBS and instance store volumes,
we could just specify a device name in the block device mapping when we
created an instance. With NVMe, the actual device is named /dev/nvme*n1,
and we have to create symlinks to the corresponding /dev/xvd* names.

This would not be needed if you could directly specify /dev/nvme*n1 in
the block device mapping, but you can't do that. You must still specify
/dev/xvd* (with or without 'ephemeral: ephemeralN', to distinguish from
EBS volumes), and deal with whatever nvme device names you end up with.
Nothing is guaranteed about how the names will be mapped, and the result
varies between instance types. For example, you can't count on instance
store volumes having lower numbers than EBS volumes, nor vice versa.

The name we specify in the block device mapping is available in the nvme
controller data for EBS volumes, but not for instance store volumes (or
in other words, ebsnvme-id doesn't work for instance store volumes). So
we can only assign a device name based on the position of the device in
the output of ``nvme list`` (and hope the order doesn't change). This is
good enough for instances with only one instance store volume, anyway.

We used to do the above for EBS volumes, albeit in a clumsy way, using
udev rules to trigger a script that created symlinks. We extend this to
support instance store volumes, and use better rules so that udev does
more of the work for us (in particular, creating the symlinks based on
the names we specify).

Signed-off-by: Abhijit Menon-Sen <ams@2ndQuadrant.com>