Configure CAMO only after BDR group join
In BDR 3.5.3, CAMO was based on synchronous replication, and configured
accordingly, esp. with synchronous_standby_names. BDR 3.5.4+ introduces
dedicated configuration (bdr.camo_{partner_of,origin_for}), but requires
these to be set after BDR is already configured, so that it can check if
the configuration is sensible. Also, these settings are PGC_POSTMASTER,
and therefore require a Postgres restart to activate.

We now install conf.d/3334-camo.conf after the BDR group join is
completed, and restart the server if required (i.e., if the
configuration file is changed). The template checks bdr_version_num
(which is defined by then, via postgres/bdr) to determine which syntax
to use. As a backwards-compatibility measure, we move 3333-bdr.conf
(which only ever contained old-CAMO configuration) to 3334-camo.conf if
the former exists but the latter doesn't.

(It doesn't seem justified to try to configure CAMOv1 before BDR and
CAMOv2 after BDR, so we just do the same new thing for both.)

With contributions from Markus.