Set «ansible_service_mgr: systemd» explicitly in minimal_setup
We assume systemd, and would blow up in more or less spectacular ways if
the system actually has something else. We don't want the service module
to do any auto-detection of some other init system and try to be clever
about which module to use (particularly because it can sometimes get it
wrong and detect sysvinit on a systemd system, as reported by Sasha).

Better to explicitly set ansible_service_mgr to avoid surprises.