Add «--install-from-source …» configure option
You can now say ``--install-from-source 2ndqpostgres pglogical3 bdr3``
to set up a cluster built from source.

This is just a convenient shortcut for the following configuration:

    cluster_vars:
        postgres_installation_method: src
        postgres_git_url: https://github.com/postgres/postgres.git
        install_from_source:
        - name: pglogical3
          git_repository_url: …
          git_repository_ref: master
          …
        - name: bdr3
          git_repository_url: …
          git_repository_ref: master
          …

If you want anything more than the default configuration (which includes
the right URLs, but not much else), you will need to edit config.yml as
before (e.g., to specify build_commands/build_environment).

Note that this option is named ``--install-from-source``, but it does
not correspond exactly to the ``install_from_source`` setting in
config.yml. The latter is not meant to be used to build Postgres from
source (which is done by setting ``postgres_installation_method: src``),
whereas the configure option does accept postgres/2ndqpostgres as a
matter of convenience.