Enable SSH agent forwarding if --install-from-source is used
We will need it to access private github repositories and the like; and
it's easier to make it work by default than to expect people to remember
to edit the inventory by hand after every provision.

It's unlikely that people will be running source builds in production
anyway, so agent forwarding is limited in practice to test clusters.