Record what needs to be done about THP, then actually do it
We used to make /etc/tpa/rc.local write the desired THP settings into
the correct places under /proc, but we don't reboot the instance by
default, so most of the time the settings remained unchanged. This
method of changing THP configuration is not 100% reliable… but not
executing the commands at all is even less effective.

We now write the commands to /etc/tpa/hugepages, invoke that script from
/etc/tpa/rc.local, but also execute the script immediately whenever its
contents change. It's the best we can do without rebooting (and if we're
going to reboot anyway, we hope the grub changes will work correctly).