Let postgres/src defer to postgres/pkg to install extra packages
Earlier, we used to hardcode python-psycopg2 and barman-cli as packages
that needed to be installed when building from source, but since we need
to use an OS-specific package for psycopg2, it makes sense now to manage
the list in postgres/pkg and defer to it to do the right thing.

It's possible that we need to distinguish between packages that depend
on Postgres being installed and those that don't, and install all of the
latter when building Postgres from source, but we can revisit this when
there are more than two packages involved, or some other compelling
reason to install additional packages for a source build.