Revise postgres_configure_opts in postgres/src
If you want to append some argument to the ./configure invocation, you
can now set postgres_extra_configure_opts without interfering with (or
having to duplicate) the builtin postgres_configure_opts; but if you
want to control all of the configure arguments, you can override the
latter as before. This change is backwards-compatible with existing
clusters.

We add --enable-tap-tests to the default postgres_configure_opts because
we want to be able to run installcheck against the same source later, so
it makes sense to do this straightaway.