Add extra dependencies for --enable-tap-tests
We need the IPC::Run and Test::More (and Time::HiRes, but that's in
core and available by default already) modules to run TAP tests. On
Debian, Test::More is available, and we can install libipc-run-perl.

On RH, Test::More is available in perl-Test-Simple, and IPC::Run is…
well, not available anywhere that I can find.