Adjust to new location of camo_client
Unfortunately, this will only work for newly-configured clusters, and
won't do anything to help existing clusters with the wrong command in
config.yml already.

It would be a little more complicated than justified to detect the right
location and adjust the command automatically to maintain compatibility
with older versions of the BDR source.