Perform bdr group join separately for instances with defined upstream
Instances that have an upstream defined should use that upstream's dsn
as the join_target_dsn when they call bdr.join_node_group(), but they
can do so only after the upstream has itself joined the group. So we
separate the process into two phases: one for instances without any
upstream, and once for the remainder.

This is important for logical standby instances, which replicate changes
only from the upstream, in order to save bandwidth, so it's necessary to
configure it correctly.

This means we don't support an instance that has an upstream defined,
but is itself defined as the upstream of another instance, although in
theory such a configuration could be sensible in future.