Introduce include-able test failures
Now any test can cause an instance or instances to fail in a particular
way by doing something like this:

    - include_role: name=test tasks_from=failures/something.yml
      vars:
        parameter: value
        …
      when:
        inventory_hostname in [some, instances] and
        …other conditions…

Initially, we define three failures:

1. powerfail: causes an instance to reboot immediately (alt-sysrq-b) and
   waits until it comes back.
2. partition: simulates a network failure for partition_timeout seconds
   on an instance; ssh traffic continues to work, so you can dispatch
   other tasks to the instance.
3. none: nothing happens.

In general, the failure is created in such a way that tasks can continue
to be executed on the degraded instance, but this depends on the nature
of the failure (e.g., an instance can't execute tasks during a power
failure).

This commit only introduces the concept of failures. Although they have
been verified to work, they are not yet used in any tests.